Split files by types
