+++ ContosoUniversity.Domain/Student/Grade.cs
@@ -1,0 +1,12 @@
+namespace ContosoUniversity.Domain.Student
+{
+    public enum Grade
+    {
+        Undefined,
+        A,
+        B,
+        C,
+        D,
+        F
+    }
+}
