+++ ContosoUniversity.Domain/Student/Student.cs
@@ -25,52 +25,4 @@ namespace ContosoUniversity.Domain.Student
         public ICollection<Enrollment> Enrollments { get; }
         public Guid EntityId { get; }
     }
-    
-    public struct Enrollment
-    {
-        private readonly Guid _courseId;
-        
-        public Enrollment(Guid courseId, Grade grade) : this()
-        {
-            CourseId = courseId;
-            Grade = grade;
-        }
-
-        public Grade Grade { get; }
-
-        public Guid CourseId
-        {
-            get => _courseId;
-            private init
-            {
-                if (value == default)
-                    throw new ArgumentException(
-                        "Course Id cannot be of default value.");
-
-                _courseId = value;
-            }
-        }
-    }
-    
-    public enum Grade
-    {
-        Undefined,
-        A,
-        B,
-        C,
-        D,
-        F
-    }
-
-    public static class GradeExtensions
-    {
-        public static string ToDisplayString(this Grade grade)
-        {
-            return grade switch
-            {
-                Grade.Undefined => "No grade",
-                var regular => regular.ToString()
-            };
-        }
-    }
 }
