+++ ContosoUniversity/ViewModels/DomainExtensions.cs
@@ -1,0 +1,16 @@
+namespace ContosoUniversity.ViewModels
+{
+    using Domain.Student;
+
+    public static class DomainExtensions
+    {
+        public static string ToDisplayString(this Grade grade)
+        {
+            return grade switch
+            {
+                Grade.Undefined => "No grade",
+                var regular => regular.ToString()
+            };
+        }
+    }
+}
